+++ cat-home/src/main/java/com/dianping/cat/report/page/dependency/service/DependencyReportService.java
@@ -52,7 +52,7 @@ public class DependencyReportService extends AbstractReportService<DependencyRep
 		throw new UnsupportedOperationException("Dependency report don't support daily report");
 	}
 
-	private DependencyReport queryFromHourlyBinary(int id, Date period, String domain) {
+	private DependencyReport queryFromHourlyBinary(long id, Date period, String domain) {
 		HourlyReportContent content = m_hourlyReportContentDao
 								.findByPK(id, period);
