+++ cat-home/src/main/java/com/dianping/cat/report/page/statistics/service/HeavyReportService.java
@@ -92,7 +92,7 @@ public class HeavyReportService extends AbstractReportService<HeavyReport> {
 		}
 	}
 
-	private HeavyReport queryFromHourlyBinary(int id, Date period, String domain) {
+	private HeavyReport queryFromHourlyBinary(long id, Date period, String domain) {
 		HourlyReportContent content = m_hourlyReportContentDao
 								.findByPK(id, period);
