+++ cat-home/src/main/java/com/dianping/cat/report/task/reload/AbstractReportReloader.java
@@ -48,7 +48,7 @@ public abstract class AbstractReportReloader implements ReportReloader {
 			HourlyReport report = entity.getReport();
 			m_hourlyReportDao.insert(report);
 
-			int id = report.getId();
+			long id = report.getId();
 			HourlyReportContent proto = m_hourlyReportContentDao.createLocal();
 
 			proto.setReportId(id);
